requirements.txt
fastapi 
uvicorn 
sentence-transformers 
faiss-cpu 
google-generativeai 
python-docx 
PyMuPDF 
python-dotenv
streamlit